requirements.txt
streamlit
python-docx
pandas
Pillow